Renamed broadcast_stream_scenario -> shared_stream_...
