+++ Benchmarks/src/main/java/io/scalecube/gateway/benchmarks/distributed/websocket/DistributedSharedStreamBenchmark.java
@@ -1,11 +1,11 @@
-package io.scalecube.gateway.benchmarks.distributed.rsocket;
+package io.scalecube.gateway.benchmarks.distributed.websocket;
 
-import io.scalecube.gateway.benchmarks.BroadcastStreamScenario;
+import io.scalecube.gateway.benchmarks.SharedStreamScenario;
 import io.scalecube.gateway.benchmarks.distributed.DistributedBenchmarkState;
 import io.scalecube.gateway.clientsdk.Client;
 import io.scalecube.gateway.clientsdk.ClientSettings;
 
-public class DistributedBroadcastStreamBenchmark {
+public class DistributedSharedStreamBenchmark {
 
   /**
    * Main runner.
@@ -13,14 +13,14 @@ public class DistributedBroadcastStreamBenchmark {
    * @param args program arguments
    */
   public static void main(String[] args) {
-    BroadcastStreamScenario.runWith(
+    SharedStreamScenario.runWith(
         args,
         benchmarkSettings ->
             new DistributedBenchmarkState(
                 benchmarkSettings,
-                "rsws",
+                "ws",
                 (address, loopResources) ->
-                    Client.onRSocket(
+                    Client.onWebsocket(
                         ClientSettings.builder()
                             .address(address)
                             .loopResources(loopResources)
