+++ Benchmarks/src/main/java/io/scalecube/gateway/benchmarks/remote/rsocket/RemoteSharedStreamBenchmark.java
@@ -2,12 +2,12 @@ package io.scalecube.gateway.benchmarks.remote.rsocket;
 
 import static io.scalecube.gateway.benchmarks.remote.RemoteBenchmarkState.RS_PORT;
 
-import io.scalecube.gateway.benchmarks.BroadcastStreamScenario;
+import io.scalecube.gateway.benchmarks.SharedStreamScenario;
 import io.scalecube.gateway.benchmarks.remote.RemoteBenchmarkState;
 import io.scalecube.gateway.clientsdk.Client;
 import io.scalecube.gateway.clientsdk.ClientSettings;
 
-public class RemoteBroadcastStreamBenchmark {
+public class RemoteSharedStreamBenchmark {
 
   /**
    * Main runner.
@@ -15,7 +15,7 @@ public class RemoteBroadcastStreamBenchmark {
    * @param args program arguments
    */
   public static void main(String[] args) {
-    BroadcastStreamScenario.runWith(
+    SharedStreamScenario.runWith(
         args,
         benchmarkSettings ->
             new RemoteBenchmarkState(
