+++ Benchmarks/src/main/java/io/scalecube/gateway/benchmarks/standalone/rsocket/StandaloneSharedStreamBenchmark.java
@@ -1,11 +1,11 @@
-package io.scalecube.gateway.benchmarks.standalone.websocket;
+package io.scalecube.gateway.benchmarks.standalone.rsocket;
 
-import io.scalecube.gateway.benchmarks.BroadcastStreamScenario;
+import io.scalecube.gateway.benchmarks.SharedStreamScenario;
 import io.scalecube.gateway.benchmarks.standalone.StandaloneBenchmarkState;
 import io.scalecube.gateway.clientsdk.Client;
 import io.scalecube.gateway.clientsdk.ClientSettings;
 
-public class StandaloneBroadcastStreamBenchmark {
+public class StandaloneSharedStreamBenchmark {
 
   /**
    * Main runner.
@@ -13,14 +13,14 @@ public class StandaloneBroadcastStreamBenchmark {
    * @param args program arguments
    */
   public static void main(String[] args) {
-    BroadcastStreamScenario.runWith(
+    SharedStreamScenario.runWith(
         args,
         benchmarkSettings ->
             new StandaloneBenchmarkState(
                 benchmarkSettings,
-                "ws",
+                "rsws",
                 (address, loopResources) ->
-                    Client.onWebsocket(
+                    Client.onRSocket(
                         ClientSettings.builder()
                             .address(address)
                             .loopResources(loopResources)
